#4626 - moved SupportedEncodings to deegree-services-commons
