+++ deegree-services/deegree-services-commons/src/main/java/org/deegree/services/encoding/UnlimitedSupportedEncodings.java
@@ -33,19 +33,17 @@
 
  e-mail: info@deegree.org
  ----------------------------------------------------------------------------*/
-package org.deegree.services.wfs.encoding;
-
-import org.deegree.protocol.wfs.WFSRequestType;
+package org.deegree.services.encoding;
 
 /**
  * {@link SupportedEncodings} implementation supporting all encodings.
  * 
  * @author <a href="mailto:goltz@lat-lon.de">Lyn Goltz</a>
  */
-public class UnlimitedSupportedEncodings implements SupportedEncodings {
+public class UnlimitedSupportedEncodings<E extends Enum> implements SupportedEncodings<E> {
 
     @Override
-    public boolean isEncodingSupported( WFSRequestType requestType, String encoding ) {
+    public boolean isEncodingSupported( E requestType, String encoding ) {
         return true;
     }
